b-q-i: Refactor the readline logic to catch multi-line messages.
A design shortcoming of the previous b-q-i was that it would always
process the stdout of the server line-by-line. Now, clean up and refactor
the io processor to allow forward buffering of lines.

Note: it is a big *assumption* that these lines all belong to the same
message. We try to flush the buffers if we see another logger-like line
or we have stdout/stderr switch. Mostly true then, we keep the assumption.