b-q-i: Implement that todo and kill the server if it doesn't terminate
At python 2.6.5, with document_ftp or some other servers, the server won't
clearly exit. We know that this is a bug, but we can't just hold the whole
buildbot process. So, try to follow the rules and carefully kill the server,
after allowing it enough time (10 sec is in the server's code) to term.