mandriva: allow modulize.py to have overriden dependency info
2 commits ago, we have supported the 'external_dependencies' style of
declaring dependencies. However, this requires that the files in question
are properly installed through RPMs. This may break some builds, because
we shall not have all these modules as build dependencies.

In case, for example, we don't want to have python-ldap installed for
the 'users_ldap' module of OpenERP, we can override with the following
line in "~/.openerp/modulize-overrides.conf" :

[python]
ldap = python-ldap